+++ apps/devportal/src/components/navigation/InPageNav.tsx
@@ -1,51 +1,18 @@
-import { Heading, List, ListIcon, ListItem, Wrap } from '@chakra-ui/react';
-import { mdiCircleSmall } from '@mdi/js';
+import { Heading, List, ListItem, Wrap } from '@chakra-ui/react';
+import { ContentHeading } from '@lib/interfaces/contentheading';
 import Link from 'next/link';
 import { title } from 'process';
-import { useEffect, useState } from 'react';
-import { ContentHeading } from '../../lib/interfaces/contentheading';
+import useInPageNavigation from '../hooks/useInPageNavigation';
 
 type InPageNavProps = {
   titles: ContentHeading[];
 };
 
 const InPageNav = ({ titles }: InPageNavProps): JSX.Element => {
-  const [links, setLinks] = useState(
-    titles.map((text) => ({
-      text: text.title,
-      href: `#${text.id}`,
-      active: false,
-    }))
-  );
-
-  useEffect(() => {
-    const observer = new IntersectionObserver(
-      (entries) => {
-        entries.forEach((entry) => {
-          if (entry.isIntersecting) {
-            const id = entry.target.getAttribute('id');
-            setLinks((prevLinks) => prevLinks.map((link) => (link.href === `#${id}` ? { ...link, active: true } : { ...link, active: false })));
-          }
-        });
-      },
-      { threshold: 1, rootMargin: '0px 0px -90% 0px' }
-    );
-
-    titles.forEach((title) => {
-      const element = document.getElementById(title.id);
-      if (element) observer.observe(element);
-    });
-
-    return () => {
-      titles.forEach((title) => {
-        const element = document.getElementById(title.id);
-        if (element) observer.unobserve(element);
-      });
-    };
-  }, [titles]);
+  const links = useInPageNavigation(titles, true);
 
   return (
-    <Wrap as={'nav'} direction="column" mt={{ base: 0, md: 10 }} mr={4} p={{ base: 2, md: 0 }} width={'2xs'} hideBelow={'xl'}>
+    <Wrap as={'nav'} direction="column" mt={{ base: 0, md: 10 }} mr={0} p={{ base: 2, md: 0 }} width={'2xs'} hideBelow={'xl'}>
       {title && (
         <Heading variant={'section'} size={'sm'} mb={{ base: 0, md: 2 }}>
           Table of contents
@@ -54,10 +21,7 @@ const InPageNav = ({ titles }: InPageNavProps): JSX.Element => {
       <List spacing="1">
         {links.map((link, i) => {
           return (
-            <ListItem display={'flex'} paddingLeft={0} key={i} fontWeight={link.active ? 'bold' : 'normal'}>
-              <ListIcon boxSize={6} mr={0}>
-                <path d={mdiCircleSmall} />
-              </ListIcon>
+            <ListItem display={'flex'} paddingLeft={0} key={i} fontWeight={link.active ? 'bold' : 'normal'} borderLeft={link.active ? '4px solid' : 'none'} borderColor={'primary-fg'} pl={link.active ? 3 : 4}>
               <Link href={link.href} key={i} title={link.text}>
                 {link.text}
               </Link>
